requirements.txt
tensorflow==2.14.0
keras==2.14.0
opencv-python==4.8.0.74
numpy==1.24.3
pyserial==3.5
matplotlib==3.7.2
scikit-learn==1.3.0
pillow==10.0.0
tqdm==4.66.1 